+++ devices/common/pci/bandit.cpp
@@ -144,7 +144,7 @@ uint32_t BanditHost::read(uint32_t rgn_start, uint32_t offset, int size)
         int bus_num, dev_num, fun_num;
         uint8_t reg_offs;
         AccessDetails details;
-        PCIDevice *device;
+        PCIBase *device;
         cfg_setup(offset, size, bus_num, dev_num, fun_num, reg_offs, details, device);
         details.flags |= PCI_CONFIG_READ;
         if (device) {
@@ -168,7 +168,7 @@ void BanditHost::write(uint32_t rgn_start, uint32_t offset, uint32_t value, int 
         int bus_num, dev_num, fun_num;
         uint8_t reg_offs;
         AccessDetails details;
-        PCIDevice *device;
+        PCIBase *device;
         cfg_setup(offset, size, bus_num, dev_num, fun_num, reg_offs, details, device);
         details.flags |= PCI_CONFIG_WRITE;
         if (device) {
@@ -196,7 +196,7 @@ void BanditHost::write(uint32_t rgn_start, uint32_t offset, uint32_t value, int 
 
 inline void BanditHost::cfg_setup(uint32_t offset, int size, int &bus_num,
                                   int &dev_num, int &fun_num, uint8_t &reg_offs,
-                                  AccessDetails &details, PCIDevice *&device)
+                                  AccessDetails &details, PCIBase *&device)
 {
     device = NULL;
     details.size = size;
