+++ devices/common/pci/bandit.h
@@ -82,7 +82,7 @@ protected:
 private:
     void cfg_setup(uint32_t offset, int size, int &bus_num, int &dev_num,
                    int &fun_num, uint8_t &reg_offs, AccessDetails &details,
-                   PCIDevice *&device);
+                   PCIBase *&device);
 };
 
 /*
