+++ devices/common/pci/pcibridge.cpp
@@ -22,16 +22,8 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #include <devices/common/pci/pcibridge.h>
 #include <memaccess.h>
 
-PCIBridge::PCIBridge(std::string name) : PCIDevice(name)
+PCIBridge::PCIBridge(std::string name) : PCIBridgeBase(name, PCI_HEADER_TYPE_1, 2)
 {
-    this->num_bars = 2;
-    this->hdr_type = 1;
-
-    this->pci_rd_primary_bus        = [this]() { return this->primary_bus; };
-    this->pci_rd_secondary_bus      = [this]() { return this->secondary_bus; };
-    this->pci_rd_subordinate_bus    = [this]() { return this->subordinate_bus; };
-    this->pci_rd_sec_latency_timer  = [this]() { return this->sec_latency_timer; };
-    this->pci_rd_sec_status         = [this]() { return this->sec_status; };
     this->pci_rd_memory_base        = [this]() { return this->memory_base; };
     this->pci_rd_memory_limit       = [this]() { return this->memory_limit; };
     this->pci_rd_io_base            = [this]() { return this->io_base; };
@@ -42,18 +34,6 @@ PCIBridge::PCIBridge(std::string name) : PCIDevice(name)
     this->pci_rd_io_limit_upper16   = [this]() { return this->io_limit_upper16; };
     this->pci_rd_pref_base_upper32  = [this]() { return this->pref_base_upper32; };
     this->pci_rd_pref_limit_upper32 = [this]() { return this->pref_limit_upper32; };
-    this->pci_rd_bridge_control     = [this]() { return this->bridge_control; };
-
-    this->pci_wr_primary_bus        = [this](uint8_t  val) { this->primary_bus = val; };
-    this->pci_wr_secondary_bus      = [this](uint8_t  val) { this->secondary_bus = val; };
-    this->pci_wr_subordinate_bus    = [this](uint8_t  val) { this->subordinate_bus = val; };
-
-    this->pci_wr_sec_latency_timer  = [this](uint8_t  val) {
-        this->sec_latency_timer = (this->sec_latency_timer & ~this->sec_latency_timer_cfg) |
-                                  (val & this->sec_latency_timer_cfg);
-    };
-
-    this->pci_wr_sec_status  = [this](uint16_t val) {};
 
     this->pci_wr_memory_base = [this](uint16_t val) {
         this->memory_base    = (val & this->memory_cfg) | (this->memory_cfg & 15);
@@ -112,33 +92,14 @@ PCIBridge::PCIBridge(std::string name) : PCIDevice(name)
         this->pref_mem_limit_64 = (((uint64_t)this->pref_limit_upper32 << 32) |
             ((this->pref_mem_limit & 0xfff0) << 16)) + 0x100000;
     };
-
-    this->pci_wr_bridge_control = [this](uint16_t val) { this->bridge_control = val; };
 };
-
-bool PCIBridge::pci_register_mmio_region(uint32_t start_addr, uint32_t size, PCIDevice* obj)
-{
-    // FIXME: constrain region to memory range
-    return this->host_instance->pci_register_mmio_region(start_addr, size, obj);
-}
-
-bool PCIBridge::pci_unregister_mmio_region(uint32_t start_addr, uint32_t size, PCIDevice* obj)
-{
-    return this->host_instance->pci_unregister_mmio_region(start_addr, size, obj);
-}
 
 uint32_t PCIBridge::pci_cfg_read(uint32_t reg_offs, AccessDetails &details)
 {
-    if (reg_offs < 0x18) {
-        return PCIDevice::pci_cfg_read(reg_offs, details);
-    }
-
     switch (reg_offs) {
-    case PCI_CFG_PRIMARY_BUS:
-        return (this->pci_rd_sec_latency_timer() << 24) |
-               (this->pci_rd_subordinate_bus() << 16)   |
-               (this->pci_rd_secondary_bus() << 8)      |
-               (this->pci_rd_primary_bus());
+    case PCI_CFG_BAR0:
+    case PCI_CFG_BAR1:
+        return this->bars[(reg_offs - 0x10) >> 2];
     case PCI_CFG_IO_BASE:
         return (this->pci_rd_sec_status() << 16) |
                (this->pci_rd_io_limit() << 8) | (this->pci_rd_io_base());
@@ -152,29 +113,21 @@ uint32_t PCIBridge::pci_cfg_read(uint32_t reg_offs, AccessDetails &details)
         return this->pci_rd_pref_limit_upper32();
     case PCI_CFG_IO_BASE_UPPER16:
         return (this->pci_rd_io_limit_upper16() << 16) | (this->pci_rd_io_base_upper16());
-    case PCI_CFG_CAP_PTR:
+    case PCI_CFG_DWORD_13:
         return cap_ptr;
     case PCI_CFG_BRIDGE_ROM_ADDRESS:
         return exp_rom_bar;
-    case PCI_CFG_INTERRUPT_LINE:
-        return (this->pci_rd_bridge_control() << 16) | (irq_pin << 8) | irq_line;
+    default:
+        return PCIBridgeBase::pci_cfg_read(reg_offs, details);
     }
-    LOG_READ_UNIMPLEMENTED_CONFIG_REGISTER();
-    return 0;
 }
 
 void PCIBridge::pci_cfg_write(uint32_t reg_offs, uint32_t value, AccessDetails &details)
 {
-    if (reg_offs < 0x18) {
-        return PCIDevice::pci_cfg_write(reg_offs, value, details);
-    }
-
     switch (reg_offs) {
-    case PCI_CFG_PRIMARY_BUS:
-        this->pci_wr_sec_latency_timer(value >> 24);
-        this->pci_wr_subordinate_bus(value >> 16);
-        this->pci_wr_secondary_bus(value >> 8);
-        this->pci_wr_primary_bus(value & 0xFFU);
+    case PCI_CFG_BAR0:
+    case PCI_CFG_BAR1:
+        this->set_bar_value((reg_offs - 0x10) >> 2, value);
         break;
     case PCI_CFG_IO_BASE:
         this->pci_wr_sec_status(value >> 16);
@@ -202,12 +155,9 @@ void PCIBridge::pci_cfg_write(uint32_t reg_offs, uint32_t value, AccessDetails &
     case PCI_CFG_BRIDGE_ROM_ADDRESS:
         this->pci_wr_exp_rom_bar(value);
         break;
-    case PCI_CFG_INTERRUPT_LINE:
-        this->irq_line = value >> 24;
-        this->pci_wr_bridge_control(value >> 16);
-        break;
     default:
-        LOG_WRITE_UNIMPLEMENTED_CONFIG_REGISTER();
+        PCIBridgeBase::pci_cfg_write(reg_offs, value, details);
+        break;
     }
 }
